% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createTiles}
\alias{createTiles}
\title{Create genome tiles for different bin sizes}
\usage{
createTiles(seqlengths, gaps, binSize = c(1000, 10000, 1e+05))
}
\arguments{
\item{seqlengths}{See \link[GenomicRanges]{tileGenome}.}

\item{gaps}{A GRanges object with the set of gaps to remove from
consideration.}

\item{binSize}{A vector with the length of the bins to use.}
}
\description{
Given a set of chromosomes, tile the genome for a set of bin sizes.
}
\examples{
## Create some input ranges
library('GenomicRanges')

## Get chr lengths
data(hg19Ideogram, package = 'biovizBase', envir = environment())
seqlengths <- seqlengths(hg19Ideogram)[c("chr1", "chrX")]

## Define some gaps to remove from tiling
gaps <- GRanges(c('chr1', 'chrX'), IRanges(c(2e8, 1e8), width = 1e6))

## Create the tiles
tiles <- createTiles(seqlengths, gaps, binSize = c(1e4, 1e5))
tiles
}
\author{
Leonardo Collado-Torres
}

