% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateSpace}
\alias{generateSpace}
\title{Obtain the null distribution of a set of intervals}
\usage{
generateSpace(I, nPermute = 1000, n = 1, ms = 10000,
  strand.specific = FALSE, randomize.strand = FALSE, seed = NULL)
}
\arguments{
\item{I}{A GRanges object giving intervals from which to sample from.}

\item{nPermute}{Number of permutations to run.}

\item{n}{Number of intervals to generate.}

\item{ms}{Length of intervals to generate.}

\item{strand.specific}{Whether the random intervals should have be strand
specific or not.}

\item{randomize.strand}{When \code{strand.specific=TRUE}, whether the strand
should be randomized or depend on the strand from the regions in \code{I}.}

\item{seed}{Seed to be used for reproducing the results.}
}
\description{
Given a set of regions of the genome, get random intervals for a number of
permutations which will be used to determine the null distribution.
}
\examples{
## Create some input ranges
library('GenomicRanges')
gr <- GRanges(c('chr1', 'chrX'), IRanges(c(1e6, 3e6), width = 1e5), strand =
    c('+', '-'))
gr

## Set the seed if you want to reproduce the random results
set.seed(20140728)

## Obtain space of random intervals
space <- generateSpace(gr, nPermute = 1000, n = 2, ms = c(1234, 4321))
space
}
\author{
Leonardo Collado-Torres
}

