% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randomInterval}
\alias{randomInterval}
\title{Get random intervals based on a set of regions}
\usage{
randomInterval(I, n = 1, ms = 10000, strand.specific = FALSE,
  randomize.strand = FALSE)
}
\arguments{
\item{I}{A GRanges object giving intervals from which to sample from.}

\item{n}{Number of intervals to generate}

\item{ms}{Length of intervals to generate. It can be a vector but
recycling rules will be used if \code{ms} is of a length smaller than
\code{n}.}

\item{strand.specific}{Whether the random intervals should have be strand
specific or not.}

\item{randomize.strand}{When \code{strand.specific=TRUE}, whether the strand
should be randomized or depend on the strand from the regions in \code{I}.}
}
\description{
Given a set of regions, generate genomic intervals uniformly at
random from those regions. These can be strand specific or not.
}
\examples{
## Create some input ranges
library('GenomicRanges')
gr <- GRanges(c('chr1', 'chrX'), IRanges(c(1e6, 3e6), width = 1e5), strand =
    c('+', '-'))
gr

## Set the seed if you want to reproduce the random results
set.seed(20140728)

## Obtain 4 randomg intervals
randomInterval(gr, 4)

## Strand specific without randomizing the strand
randomInterval(gr, 4, strand.specific = TRUE)

## This function is a modified version of seqbias::random.intervals
## Check the original code below:
library('seqbias')
random.intervals
## Some key differences are that randomInterval works with repeated
## seqnames and starts at the beginning of the input ranges, not at base 1.
}
\author{
Leonardo Collado-Torres
}

